requirements.txt
librosa~=0.9.2
numpy~=1.22.2
soxr~=0.3.4
tensorflow~=2.11.0
pillow~=9.3.0
scipy~=1.6.3